+++ assets/javascripts/kitten/components/buttons/button-icon/index.js
