+++ assets/javascripts/kitten/components/buttons/button-image-with-text-and-badge/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
-import COLORS from '../../constants/colors-config'
+import COLORS from '../../../constants/colors-config'
 
 const NotificationBadge = ({ notifications, badgeColor }) => (
   <span
