+++ assets/javascripts/kitten/components/buttons/button-image/index.js
