+++ assets/javascripts/kitten/components/buttons/button-question-mark-icon/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { QuestionMarkIcon } from '../../components/icons/question-mark-icon'
-import { ButtonIcon } from '../../components/buttons/button-icon'
+import { QuestionMarkIcon } from '../../../components/icons/question-mark-icon'
+import { ButtonIcon } from '../../../components/buttons/button-icon'
 
 export const ButtonQuestionMarkIcon = props => (
   <ButtonIcon
