+++ assets/javascripts/kitten/components/buttons/button/deprecated/index.js
