+++ assets/javascripts/kitten/components/buttons/button/index.js
@@ -1,0 +1,3 @@
+import { Button } from './deprecated/index.js'
+
+export { Button }
