+++ assets/javascripts/kitten/components/buttons/social-button-icon-words/test.js
@@ -3,11 +3,11 @@ import {
   FacebookButtonIconWords,
   TwitterButtonIconWords,
   LinkedinButtonIconWords,
-} from '../../components/buttons/social-button-icon-words'
-import { Button } from '../../components/buttons/button'
-import { FacebookIcon } from '../../components/icons/facebook-icon'
-import { TwitterIcon } from '../../components/icons/twitter-icon'
-import { LinkedinIcon } from '../../components/icons/linkedin-icon'
+} from '../../../components/buttons/social-button-icon-words'
+import { Button } from '../../../components/buttons/button'
+import { FacebookIcon } from '../../../components/icons/facebook-icon'
+import { TwitterIcon } from '../../../components/icons/twitter-icon'
+import { LinkedinIcon } from '../../../components/icons/linkedin-icon'
 
 describe('<FacebookButtonIconWords />', () => {
   describe('by default', () => {
