+++ assets/javascripts/kitten/components/buttons/social-button-icon/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import { ButtonIcon } from '../../components/buttons/button-icon'
-import { FacebookIcon } from '../../components/icons/facebook-icon'
-import { TwitterIcon } from '../../components/icons/twitter-icon'
-import { LinkedinIcon } from '../../components/icons/linkedin-icon'
-import { InstagramIcon } from '../../components/icons/instagram-icon'
-import { YoutubeIcon } from '../../components/icons/youtube-icon'
-import { PinterestIcon } from '../../components/icons/pinterest-icon'
+import { ButtonIcon } from '../../../components/buttons/button-icon'
+import { FacebookIcon } from '../../../components/icons/facebook-icon'
+import { TwitterIcon } from '../../../components/icons/twitter-icon'
+import { LinkedinIcon } from '../../../components/icons/linkedin-icon'
+import { InstagramIcon } from '../../../components/icons/instagram-icon'
+import { YoutubeIcon } from '../../../components/icons/youtube-icon'
+import { PinterestIcon } from '../../../components/icons/pinterest-icon'
 
 export const FacebookButtonIcon = props => (
   <ButtonIcon {...props} modifier="facebook">
