+++ assets/javascripts/kitten/components/buttons/social-button-icon/test.js
@@ -7,14 +7,14 @@ import {
   InstagramButtonIcon,
   YoutubeButtonIcon,
   PinterestButtonIcon,
-} from '../../components/buttons/social-button-icon'
-import { ButtonIcon } from '../../components/buttons/button-icon'
-import { FacebookIcon } from '../../components/icons/facebook-icon'
-import { TwitterIcon } from '../../components/icons/twitter-icon'
-import { LinkedinIcon } from '../../components/icons/linkedin-icon'
-import { InstagramIcon } from '../../components/icons/instagram-icon'
-import { YoutubeIcon } from '../../components/icons/youtube-icon'
-import { PinterestIcon } from '../../components/icons/pinterest-icon'
+} from '../../../components/buttons/social-button-icon'
+import { ButtonIcon } from '../../../components/buttons/button-icon'
+import { FacebookIcon } from '../../../components/icons/facebook-icon'
+import { TwitterIcon } from '../../../components/icons/twitter-icon'
+import { LinkedinIcon } from '../../../components/icons/linkedin-icon'
+import { InstagramIcon } from '../../../components/icons/instagram-icon'
+import { YoutubeIcon } from '../../../components/icons/youtube-icon'
+import { PinterestIcon } from '../../../components/icons/pinterest-icon'
 
 describe('<FacebookButtonIcon />', () => {
   describe('by default', () => {
